[NetworkManager] move NetworkManager-specific content to a new plugin.
As NetworkManager is being further developed as a stand-alone service,
it is likely that new output will be necessary.  Therefore, further
cluttering the existing networking plugin may be inadvisable and the
two should be seperated.

This commit simply removes any NetworkManager-specific output from the
networking plugin and moves it to it's own NetworkManager plugin.  From
there expansion on to the service can be done easily without fear of
disruption of the underlying networking plugin.
